refactor create entry again to be for outdoors
