+++ app/src/main/res/values/strings.xml
@@ -8,10 +8,7 @@
     <string name="Schedule">Schedule</string>
     <string name="loading_image_description">loading image</string>
     <string name="DropDown">DropDown</string>
-    <string name="Name">Name</string>
     <string name="Location">Location</string>
-    <string name="Plant_Type">Plant Type</string>
-    <string name="Adoption_Date">Adoption Date</string>
-    <string name="Water">Water</string>
-    <string name="Light">Light</string>
+    <string name="Seed_Type">Seed Type</string>
+    <string name="Plant_Category">Plant Category</string>
 </resources>
